fix guzzle promise handling (#393)
In certain circumstances (eg Curl handler and middlewares configured such that no additional promise chaining
occurs), our post transfer hook can receive an already-resolved Promise. Since it's already resolved, our
extra callbacks won't be applied in a timely fashion (they're instead added to a queue and will be processed
as part of a shutdown handler, if not triggered by something earlier).
The upshot of this is a lot of memory being held on to for each request, and it seems that the default config
when we use guzzle as our SDK exporter will trigger this. The spans for export are non-recording, so nothing
user-visible happens, but the non-recording spans and promise callbacks are held in memory, seemingly forever
or until shutdown.
The fix for this is to check is the promise is already resolved, and if so force our extra promise to resolve.